Use env variable instead of server route to control time plugin checking
