+++ .env
@@ -6,5 +6,6 @@ PUBLIC_HASURA_CLIENT_URL=http://localhost:8080/v1/graphql
 PUBLIC_HASURA_SERVER_URL=http://localhost:8080/v1/graphql
 PUBLIC_HASURA_WEB_SOCKET_URL=ws://localhost:8080/v1/graphql
 PUBLIC_AUTH_SSO_ENABLED=false
+PUBLIC_TIME_PLUGIN_ENABLED=true
 # VITE_HOST=localhost.jpl.nasa.gov
 # VITE_HTTPS=true
