+++ src/stores/plugins.ts
@@ -4,3 +4,4 @@ import type { Plugins } from '../types/plugin';
 /* Writeable. */
 
 export const plugins: Writable<Plugins> = writable({});
+export const pluginsLoaded: Writable<boolean> = writable(false);
